refactor: more cleanup and documentation
