+++ tests/data/unit_tests/test_analyzer/test_case_statement_parsing.sql
