+++ tests/data/unit_tests/test_analyzer/test_cte_parsing.sql
