+++ tests/data/unit_tests/test_analyzer/test_simple_query_parsing.sql
