+++ tests/data/unit_tests/test_merger/test_merge_operators_before_children.sql
