+++ tests/data/unit_tests/test_splitter/test_comment_split_impact_on_open_brackets.sql
