+++ tests/data/unit_tests/test_splitter/test_simple_comment_split.sql
@@ -17,3 +17,4 @@ from
     my_really_long_data_source
 where
     true
+
